began work updating sign-in to include error messages
