+++ app/assets/stylesheets/layouts/_home.scss
@@ -10,7 +10,7 @@
     width: 100%;
   }
 }
-.home__signup {
+.home__signIn {
   width: 100%;
   max-width: 350px;
 }
