+++ app/views/api/errors.json.jbuilder
@@ -1,2 +1,4 @@
 json.object_type 'errors'
-json.errors @errors
+json.errors @errors.each do |k,v|
+  json.set! k, v
+end
