+++ app/views/home/index.html.haml
@@ -4,7 +4,7 @@
 %section.container-fluid.container-table
   .row.vertical-center-row
     .col-xs-10.col-xs-offset-1
-      .home__signup.center-block
+      .home__signIn.center-block
         %image{ src:image_path('HW-logo.png'), class: 'img-responsive center-block logo__large'}
         %h1.home__title.text-center
           Habitual Wealth
