+++ package.json
@@ -36,6 +36,7 @@
     "history": "^1.13.1",
     "react": "^0.14.3",
     "react-dom": "^0.14.3",
+    "react-redux": "^4.0.0",
     "react-router": "^1.0.0",
     "reactify": "^1.1.1",
     "redux": "^3.0.4",
